Moves connectionManagement to src/db
